+++ src/ReportBuilder.php
@@ -1,0 +1,13 @@
+<?php
+
+/** @noinspection PhpUndefinedMethodInspection */
+
+namespace ACTTraining\QueryBuilder;
+
+use ACTTraining\QueryBuilder\Support\Concerns\WithColumnsBuilder;
+
+abstract class ReportBuilder extends TableBuilder
+{
+    use WithColumnsBuilder;
+
+}
