+++ src/Support/Concerns/WithColumnsBuilder.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace ACTTraining\QueryBuilder\Support\Concerns;
+
+trait WithColumnsBuilder
+{
+    public function configuredColumns(): array
+    {
+        return [];
+    }
+
+    public function buildColumns(): array
+    {
+        return [];
+    }
+
+}
